refactor(cli): split cli_args.rs into focused submodules
The single cli_args.rs file had grown to ~1500 lines and mixed several
unrelated concerns, making it a frequent source of merge conflicts. Split
it into focused submodules under cli_args/, leaving cli_args.rs as a thin
module index:

- cli_command.rs: the CliArgs and CliCommand argument types
- dispatch.rs:    the impl CliArgs command-dispatch logic
- pipelines.rs:   the install/deploy/dedupe/prune reporter-generic pipelines
- package_manager.rs: packageManager / devEngines parsing helpers
- reporter.rs:    ReporterType and its emit/configure helpers

No behavior change; the move is mechanical. Subcommand modules that adding
a new command still touches (the enum and the dispatch match) now live in
their own files, so unrelated changes to pipelines or package-manager
parsing no longer conflict with them.
